refactor specialized
