+++ crates/cubecl-matmul/src/components/global/specialization/config.rs
@@ -1,0 +1,173 @@
+use crate::{
+    components::{
+        InputIdent,
+        global::{MaxLoaders, specialization::roles::PlaneRoles},
+    },
+    gcd,
+};
+
+/// Configuration for how each input tensor (Lhs and Rhs) is loaded,
+/// specifying the plane roles responsible for loading them.
+#[derive(Default, Copy, Clone, Debug, Hash, PartialEq, Eq)]
+pub struct LoadSpecializationConfig {
+    /// Load strategy for the Lhs tensor.
+    pub lhs: SpecializationTensorConfig,
+    /// Load strategy for the Rhs tensor.
+    pub rhs: SpecializationTensorConfig,
+}
+
+/// Determines which types of planes are responsible for loading a tensor.
+///
+/// TODO: maybe we want a "MainPlusExtra" variant that uses main flow planes and load-only planes
+/// for the same tensor
+#[derive(Default, Copy, Clone, Debug, Hash, PartialEq, Eq)]
+pub enum SpecializationTensorConfig {
+    /// The tensor is loaded exclusively by planes that participate in the main computation flow.
+    #[default]
+    MainFlowOnly,
+
+    /// The tensor is loaded exclusively by planes dedicated to loading (load-only planes),
+    /// which do not participate in computation.
+    LoadFlowOnly,
+}
+
+impl LoadSpecializationConfig {
+    /// Whether there is specialization in the algorithm
+    pub fn has_specialization(&self) -> bool {
+        self.lhs.has_specialization() || self.rhs.has_specialization()
+    }
+}
+
+impl SpecializationTensorConfig {
+    /// Whether there is specialization for the tensor
+    pub fn has_specialization(&self) -> bool {
+        match self {
+            SpecializationTensorConfig::MainFlowOnly => false,
+            SpecializationTensorConfig::LoadFlowOnly => true,
+        }
+    }
+}
+
+impl LoadSpecializationConfig {
+    ///
+    pub fn to_plane_roles(&self, main_flow: u32, loader_tasks: MaxLoaders) -> PlaneRoles {
+        PlaneRoles {
+            main_flow,
+            load_only: self.find_num_load_only(main_flow, loader_tasks),
+        }
+    }
+
+    fn find_num_load_only(&self, main_flow: u32, loader_tasks: MaxLoaders) -> u32 {
+        use SpecializationTensorConfig::*;
+
+        let ideal_load_only = match (self.lhs, self.rhs) {
+            (MainFlowOnly, MainFlowOnly) => 0,
+            (MainFlowOnly, LoadFlowOnly) => loader_tasks.rhs,
+            (LoadFlowOnly, MainFlowOnly) => loader_tasks.lhs,
+            (LoadFlowOnly, LoadFlowOnly) => gcd(loader_tasks.lhs, loader_tasks.rhs),
+        };
+
+        // Don't stray too far from main_flow
+        best_divisor_close_to_reference(ideal_load_only, main_flow)
+    }
+}
+
+fn best_divisor_close_to_reference(dividible_value: u32, reference: u32) -> u32 {
+    let mut best = 1;
+    let mut best_dist = reference.abs_diff(1);
+
+    for d in 1..=dividible_value {
+        if dividible_value % d == 0 {
+            let dist = reference.abs_diff(d);
+            if dist < best_dist || (dist == best_dist && d > best) {
+                best = d;
+                best_dist = dist;
+            }
+        }
+    }
+
+    best
+}
+
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
+pub enum LoadingSides {
+    // Load both Lhs and Rhs
+    Both,
+    // Load Lhs only
+    Lhs,
+    // Load Rhs only
+    Rhs,
+    // Don't perform loading
+    None,
+}
+
+impl LoadingSides {
+    pub fn includes_lhs(&self) -> bool {
+        self.includes(InputIdent::Lhs)
+    }
+
+    pub fn includes_rhs(&self) -> bool {
+        self.includes(InputIdent::Rhs)
+    }
+
+    pub fn includes(&self, ident: InputIdent) -> bool {
+        matches!(
+            (self, ident),
+            (LoadingSides::Both, _)
+                | (LoadingSides::Lhs, InputIdent::Lhs)
+                | (LoadingSides::Rhs, InputIdent::Rhs)
+        )
+    }
+}
+
+#[derive(Copy, Clone, Debug, Hash, PartialEq, Eq)]
+pub struct SpecializedLoadingSides {
+    pub main_flow: LoadingSides,
+    pub load_only: LoadingSides,
+}
+
+impl SpecializedLoadingSides {
+    pub fn num_loading_planes(
+        &self,
+        specialized: bool,
+        ident: InputIdent,
+        plane_roles: PlaneRoles,
+    ) -> u32 {
+        if specialized {
+            let mut num_loading_planes = 0;
+            if self.main_flow.includes(ident) {
+                num_loading_planes += plane_roles.main_flow;
+            }
+            if self.load_only.includes(ident) {
+                num_loading_planes += plane_roles.load_only;
+            }
+            num_loading_planes
+        } else {
+            plane_roles.main_flow
+        }
+    }
+}
+
+impl From<LoadSpecializationConfig> for SpecializedLoadingSides {
+    fn from(lsc: LoadSpecializationConfig) -> Self {
+        use SpecializationTensorConfig::*;
+        match (lsc.lhs, lsc.rhs) {
+            (MainFlowOnly, MainFlowOnly) => SpecializedLoadingSides {
+                main_flow: LoadingSides::Both,
+                load_only: LoadingSides::None,
+            },
+            (MainFlowOnly, LoadFlowOnly) => SpecializedLoadingSides {
+                main_flow: LoadingSides::Lhs,
+                load_only: LoadingSides::Rhs,
+            },
+            (LoadFlowOnly, MainFlowOnly) => SpecializedLoadingSides {
+                main_flow: LoadingSides::Rhs,
+                load_only: LoadingSides::Lhs,
+            },
+            (LoadFlowOnly, LoadFlowOnly) => SpecializedLoadingSides {
+                main_flow: LoadingSides::None,
+                load_only: LoadingSides::Both,
+            },
+        }
+    }
+}
