+++ crates/cubecl-matmul/src/components/global/specialization/mod.rs
@@ -1,0 +1,11 @@
+mod config;
+mod roles;
+mod specializer;
+
+pub use config::{
+    LoadSpecializationConfig, LoadingSides, SpecializationTensorConfig, SpecializedLoadingSides,
+};
+pub use roles::{PlaneRoleConfig, RoleRule, RoleRuleConfig};
+pub use specializer::{Specializer, SpecializerKind};
+
+// PlaneRoleConfig, SpecializedLoadingSides,
