+++ crates/cubecl-matmul/src/components/tiling_scheme.rs
@@ -2,6 +2,8 @@ use super::Ident;
 use super::size::{GlobalPartitionSize, MatmulDim, PartitionSize, StageSize, TileSize};
 
 #[derive(Debug, Copy, Clone, Hash, PartialEq, Eq)]
+/// Complete tiling configuration for a matmul.
+/// Encodes all structural information needed to compute tiling shapes and counts.
 pub struct TilingScheme {
     pub tile_size: TileSize,
     pub partition_size: PartitionSize,
@@ -10,12 +12,14 @@ pub struct TilingScheme {
 }
 
 impl TilingScheme {
+    /// Start a builder for TilingScheme
     pub fn builder() -> TilingSchemeBuilder {
         TilingSchemeBuilder::default()
     }
 }
 
 #[derive(Debug, Default)]
+/// Builder for [`TilingScheme`]. Allows step-by-step configuration.
 pub struct TilingSchemeBuilder {
     tile_size: Option<TileSize>,
     partition_size: Option<PartitionSize>,
@@ -24,22 +28,30 @@ pub struct TilingSchemeBuilder {
 }
 
 impl TilingSchemeBuilder {
+    /// Specify tile size for tiling scheme
     pub fn with_tile_size(mut self, tile_size: TileSize) -> Self {
         self.tile_size = Some(tile_size);
         self
     }
 
+    /// Specify partition size for tiling scheme
     pub fn with_partition_size(mut self, partition_size: PartitionSize) -> Self {
         self.partition_size = Some(partition_size);
         self
     }
 
+    /// Specify stage size for tiling scheme
+    ///
+    /// Only stage size k = 1 is supported
     pub fn with_stage_size(mut self, stage_size: StageSize) -> Self {
         assert!(stage_size.k == 1, "Stage size k > 1 is not supported");
         self.stage_size = Some(stage_size);
         self
     }
 
+    /// Optional: specify global partition size for tiling scheme
+    ///
+    /// If not specified, will default to (1, 1, 1)
     pub fn with_global_partition_size(
         mut self,
         global_partition_size: GlobalPartitionSize,
@@ -48,6 +60,7 @@ impl TilingSchemeBuilder {
         self
     }
 
+    /// Finish building
     pub fn build(self) -> Result<TilingScheme, &'static str> {
         Ok(TilingScheme {
             tile_size: self.tile_size.ok_or("Missing tile_size")?,
@@ -61,7 +74,7 @@ impl TilingSchemeBuilder {
 }
 
 #[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub enum TilingLevel {
+enum TilingLevel {
     GlobalPartition,
     Stage,
     StagePartition,
